Remove deprecated methods
Instigated by inability to call -[NSManagedObjectContext save:] from Swift because of deprecated MagicalRecord -save method (the former is imported as a throws function with the same signature as the latter, and the latter apparently takes precedence in this case).

There were some test failures before and there are some test failures now, but they don't seem too worrying.